Read Stata DTA files (`.dta`) files

This is a wrapper around haven::read_dta with some exception handling.

read_dta(file, id = NULL, filename = NULL, doi = NULL, .name_repair = "unique")

Arguments

file
A STATA file.

id
An identifier of the tibble, if omitted, defaults to the file name.

filename
An import file name.

doi
An optional document object identifier.

.name_repair
Defaults to "unique" See tibble::as_tibble for details.

user_na
Should user-defined na_values be imported? Defaults to TRUE.

Value

A tibble.

Variable labels are stored in the "label" attribute of each variable. It is not printed on the console, but the RStudio viewer will show it.

`write_sav()` returns the input `data` invisibly.

Details

`read_rda()` reads both `.dta` files.

The funcion is not yet tested.

See also

Examples

# \donttest{
path <- system.file("examples", "iris.dta", package = "haven")
read_dta(path)

#> Error in read_dta(path): could not find function "read_dta"

# }