Suggest Variable Names The function harmonizes the variable names of surveys (of class survey) that are imported from an external file as a wave.

Suggest Variable Names

The function harmonizes the variable names of surveys (of class survey) that are imported from an external file as a wave.

suggest_var_names(
  metadata,
  permanent_names = NULL,
  survey_program = NULL,
  case = "snake"
)

Arguments

metadata
A metadata table created by metadata_create and binded together for all surveys in waves.

permanent_names
A character vector of names to keep.

survey_program
If permanent_names = NULL then suggest_permanent_names is called with this parameter, unless it is also NULL

case
Unless it is set to NULL it will standardize the suggested variable name with to_any_case. The default is "snake".

Value

A metadata tibble augmented with $var_name_suggested

See also

Examples

examples_dir <- system.file("examples", package = "retroharmonize")
survey_list <- dir(examples_dir)[grepl("\\.rds", dir(examples_dir))]

example_surveys <- read_surveys(
  file.path(examples_dir, survey_list), 
  save_to_rds = FALSE)
metadata <- lapply ( X = example_surveys, FUN = metadata_create )
metadata <- do.call(rbind, metadata)

utils::head(
  suggest_var_names(metadata, survey_program = "eurobarometer" )
  )

#>     filename     id var_name_orig     class_orig
#> 1 ZA5913.rds ZA5913         rowid      character
#> 2 ZA5913.rds ZA5913           doi      character
#> 3 ZA5913.rds ZA5913       version      character
#> 4 ZA5913.rds ZA5913        uniqid        numeric
#> 5 ZA5913.rds ZA5913      isocntry      character
#> 6 ZA5913.rds ZA5913            p1 haven_labelled
#>                                        label_orig
#> 1                     unique identifier in za5913
#> 2                       digital object identifier
#> 3                  gesis archive version and date
#> 4 unique respondent id caseid by tns country code
#> 5                           country code iso 3166
#> 6                               date of interview
#>                                          labels
#> 1                                            NA
#> 2                                            NA
#> 3                                            NA
#> 4                                            NA
#> 5                                            NA
#> 6 1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12, 13, 14
#>                                    valid_labels na_labels na_range n_labels
#> 1                                            NA        NA       NA        0
#> 2                                            NA        NA       NA        0
#> 3                                            NA        NA       NA        0
#> 4                                            NA        NA       NA        0
#> 5                                            NA        NA       NA        0
#> 6 1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12, 13, 14                 NA       14
#>   n_valid_labels n_na_labels                              var_name_suggested
#> 1              0           0                                           rowid
#> 2              0           0                       digital_object_identifier
#> 3              0           0                  gesis_archive_version_and_date
#> 4              0           0 unique_respondent_id_caseid_by_tns_country_code
#> 5              0           0                                        isocntry
#> 6             14           0                               date_of_interview