Create a codebook

Create a codebook from one or more survey data files.

Usage

create_codebook(metadata = NULL, survey = NULL)

codebook_waves_create(waves)

codebook_surveys_create(survey_list)

Arguments

metadata

A metadata table created by metadata_create. Defaults to NULL.

survey

A survey data frame, defaults to NULL. If the survey is given as parameter, the metadata will be set to the metadata of this particular survey by metadata_create.

waves

A list of surveys.

survey_list

A list containing surveys of class survey.

Value

A codebook for the survey as a data frame, including the metadata, and all found SPSS-type valid or missing labels.

Details

For a list of survey waves, use codebook_waves_create. The returned codebook contains only labelled variables, i.e., numeric and character types are not included, because they do not require coding.

See also

Examples

create_codebook (
 survey = read_rds (
          system.file("examples", "ZA7576.rds",
                      package = "retroharmonize")
          )
)
#> Warning: Unknown or uninitialised column: `rowid`.
#> # A tibble: 535 × 12
#>    entry id     filename   var_name_orig var_label_orig    val_code_orig
#>    <int> <chr>  <chr>      <chr>         <chr>             <chr>        
#>  1     7 ZA7576 ZA7576.rds p1            date_of_interview 1            
#>  2     7 ZA7576 ZA7576.rds p1            date_of_interview 10           
#>  3     7 ZA7576 ZA7576.rds p1            date_of_interview 11           
#>  4     7 ZA7576 ZA7576.rds p1            date_of_interview 12           
#>  5     7 ZA7576 ZA7576.rds p1            date_of_interview 13           
#>  6     7 ZA7576 ZA7576.rds p1            date_of_interview 14           
#>  7     7 ZA7576 ZA7576.rds p1            date_of_interview 15           
#>  8     7 ZA7576 ZA7576.rds p1            date_of_interview 16           
#>  9     7 ZA7576 ZA7576.rds p1            date_of_interview 17           
#> 10     7 ZA7576 ZA7576.rds p1            date_of_interview 18           
#> # ℹ 525 more rows
#> # ℹ 6 more variables: val_label_orig <chr>, label_range <chr>,
#> #   na_range <named list>, n_labels <dbl>, n_valid_labels <dbl>,
#> #   n_na_labels <dbl>
# \donttest{
examples_dir <- system.file("examples", package = "retroharmonize")
survey_list <- dir(examples_dir)[grepl("\\.rds", dir(examples_dir))]

example_surveys <- read_surveys(
  file.path( examples_dir, survey_list), 
  save_to_rds = FALSE)     
#> Warning: Unknown or uninitialised column: `rowid`.
#> Warning: Unknown or uninitialised column: `rowid`.
#> Warning: Unknown or uninitialised column: `rowid`.

codebook_surveys_create (example_surveys)
#> # A tibble: 1,366 × 12
#>    entry id     filename   var_name_orig var_label_orig    val_code_orig
#>    <int> <chr>  <chr>      <chr>         <chr>             <chr>        
#>  1     5 ZA5913 ZA5913.rds p1            date_of_interview 1            
#>  2     5 ZA5913 ZA5913.rds p1            date_of_interview 10           
#>  3     5 ZA5913 ZA5913.rds p1            date_of_interview 11           
#>  4     5 ZA5913 ZA5913.rds p1            date_of_interview 12           
#>  5     5 ZA5913 ZA5913.rds p1            date_of_interview 13           
#>  6     5 ZA5913 ZA5913.rds p1            date_of_interview 14           
#>  7     5 ZA5913 ZA5913.rds p1            date_of_interview 2            
#>  8     5 ZA5913 ZA5913.rds p1            date_of_interview 3            
#>  9     5 ZA5913 ZA5913.rds p1            date_of_interview 4            
#> 10     5 ZA5913 ZA5913.rds p1            date_of_interview 5            
#> # ℹ 1,356 more rows
#> # ℹ 6 more variables: val_label_orig <chr>, label_range <chr>,
#> #   na_range <named list>, n_labels <dbl>, n_valid_labels <dbl>,
#> #   n_na_labels <dbl>
# }