Document survey lists

Document the key attributes surveys in a survey list.

Usage

document_surveys(survey_list)

document_waves(survey_list, var_harmonization)

Arguments

survey_list

A list of survey objects.

Value

Returns a data frame with the key attributes of the surveys in a survey list: the name of the data file, the number of rows and columns, and the size of the object as stored in memory.

Details

The earlier form document_waves is deprecated. Currently called document_surveys.

See also

Examples

examples_dir <- system.file( "examples", package = "retroharmonize")
                        
my_rds_files <- dir( examples_dir)[grepl(".rds", 
                                   dir(examples_dir))]

example_surveys <- read_surveys(file.path(examples_dir, my_rds_files))
 
documented_surveys <- document_surveys(example_surveys)

attr(documented_surveys, "original_list" )
#> [1] "example_surveys"
documented_surveys
#> # A tibble: 3 × 5
#>   id     filename    ncol  nrow object_size
#>   <chr>  <chr>      <int> <int>       <dbl>
#> 1 ZA5913 ZA5913.rds    38    35      110192
#> 2 ZA6863 ZA6863.rds    49    50      144848
#> 3 ZA7576 ZA7576.rds    56    45      165824