Merge surveys

Merge a list of surveys into a list with harmonized variable names, variable labels and survey identifiers.

Usage

merge_surveys(survey_list, var_harmonization)

merge_waves(waves, var_harmonization)

Arguments

survey_list

A list of surveys

var_harmonization

Metadata of surveys, including at least filename, var_name_orig, var_name_target, var_label.

waves

Deprecated.

Value

A list of surveys with harmonized names and variable labels.

Details

The function was called till version 0.2.0 merge_waves(), which reflects the vocabulary of Eurobarometer surveys.

See also

Examples

# \donttest{
examples_dir <- system.file("examples", package = "retroharmonize")
survey_list <- dir(examples_dir)[grepl("\\.rds", dir(examples_dir))]

example_surveys <- read_surveys(
  file.path( examples_dir, survey_list), 
  save_to_rds = FALSE)
    
metadata <- metadata_surveys_create(example_surveys)
 
require(dplyr)

to_harmonize <- metadata %>%
  filter ( var_name_orig %in% 
             c("rowid", "w1") |
             grepl("^trust", label_orig ) ) %>%
  mutate ( var_label = var_label_normalize(label_orig) ) %>%
  mutate ( var_name_target = val_label_normalize(var_label) ) %>%
  mutate ( var_name_target = ifelse(.data$var_name_orig %in% c("rowid", "w1", "wex"), 
                                    .data$var_name_orig, .data$var_name_target) )
#> Error: Problem with `filter()` input `..1`.
#> ℹ Input `..1` is `var_name_orig %in% c("rowid", "w1") | grepl("^trust", label_orig)`.
#> ✖ object 'label_orig' not found

merge_surveys ( example_surveys, to_harmonize )
#> Error in c("filename", "var_name_orig", "var_name_target", "var_label") %in%     names(var_harmonization): object 'to_harmonize' not found
# }