Merge waves

Merge a list of surveys into a list with harmonized variable names, variable labels and survey identifiers.

merge_waves(waves, var_harmonization)

Arguments

waves
A list of surveys

var_harmonization
Metadata of surveys, including at least filename, var_name_orig, var_name, var_label.

Value

A list of surveys with harmonized names and labels.

See also