Read Survey Files

Import surveys into a list. Adds filename as a constant to each element of the list.

Usage

read_surveys(import_file_names, .f = "read_rds", save_to_rds = FALSE)

Arguments

import_file_names

A vector of file names to import.

.f

A function to import the surveys with. Defaults to 'read_rds'. For SPSS files, read_spss is recommended, which is a well-parameterized version of read_spss that saves some metadata, too.

save_to_rds

Should it save the imported survey to .rds? Defaults to FALSE.

Value

A list of the surveys. Each element of the list is a data frame-like survey type object where some metadata, such as the original file name, doi identifier if present, and other information is recorded for a reproducible workflow.

Details

The functions handle exceptions with wrong filenames and not readable files. If I file cannot be read, a warning is given, and empty survey is added to the the list in the place of this file.

See also

Examples

file1 <- system.file(
    "examples", "ZA7576.rds", package = "retroharmonize")
file2 <- system.file(
    "examples", "ZA5913.rds", package = "retroharmonize")

read_surveys (c(file1,file2), .f = 'read_rds' )
#> [[1]]
#> # A tibble: 45 × 55
#>    rowid  doi    version uniqid caseid serialid isocntry        p1      p2    p3
#>  * <chr>  <chr>  <chr>    <dbl>  <dbl>    <dbl> <chr>    <dbl+lbl> <dbl+l> <dbl>
#>  1 ZA757… doi:1… 1.0.0 … 5.00e7    481     3209 ES        4 [Mond… 3 [13 …    25
#>  2 ZA757… doi:1… 1.0.0 … 1.10e8     76     8706 NL        6 [Wedn… 3 [13 …    58
#>  3 ZA757… doi:1… 1.0.0 … 1.10e8    343     8890 NL       11 [Mond… 3 [13 …    56
#>  4 ZA757… doi:1… 1.0.0 … 1.10e8    473     8989 NL        5 [Tues… 3 [13 …    62
#>  5 ZA757… doi:1… 1.0.0 … 1.10e8    493     9001 NL        8 [Frid… 4 [17 …    30
#>  6 ZA757… doi:1… 1.0.0 … 1.10e8    897     9272 NL        6 [Wedn… 3 [13 …    56
#>  7 ZA757… doi:1… 1.0.0 … 1.10e8   1041     9379 NL        5 [Tues… 3 [13 …    57
#>  8 ZA757… doi:1… 1.0.0 … 1.10e8   1192     9493 NL        6 [Wedn… 2 [8 -…    60
#>  9 ZA757… doi:1… 1.0.0 … 1.10e8   1274     9543 NL        7 [Thur… 4 [17 …    57
#> 10 ZA757… doi:1… 1.0.0 … 1.10e8   1344     9590 NL        6 [Wedn… 2 [8 -…    83
#> # … with 35 more rows, and 45 more variables: p4 <dbl+lbl>, nuts <chr+lbl>,
#> #   d7 <dbl+lbl>, d8 <dbl+lbl>, d25 <dbl+lbl>, d60 <dbl+lbl>, qa14_5 <dbl+lbl>,
#> #   qa14_3 <dbl+lbl>, qa14_2 <dbl+lbl>, qa14_4 <dbl+lbl>, qa14_1 <dbl+lbl>,
#> #   qa6a_5 <dbl+lbl>, qa6a_10 <dbl+lbl>, qa6b_2 <dbl+lbl>, qa6a_3 <dbl+lbl>,
#> #   qa6a_1 <dbl+lbl>, qa6b_4 <dbl+lbl>, qa6a_8 <dbl+lbl>, qa6a_9 <dbl+lbl>,
#> #   qa6a_4 <dbl+lbl>, qa6a_2 <dbl+lbl>, qa6b_1 <dbl+lbl>, qa6a_6 <dbl+lbl>,
#> #   qa6a_7 <dbl+lbl>, qa6a_11 <dbl+lbl>, qa6b_3 <dbl+lbl>, qd6.1 <dbl+lbl>, …
#> 
#> [[2]]
#> # A tibble: 35 × 37
#>    rowid     doi    version   uniqid isocntry       p1    p3       p4 nuts      
#>  * <chr>     <chr>  <chr>      <dbl> <chr>    <dbl+lb> <dbl> <dbl+lb> <chr+lbl> 
#>  1 ZA5913_1  doi:1… 2.0.0 (…  1.13e7 NL        8 [Tue…    27 1 [Two … NL21 [Ove…
#>  2 ZA5913_2  doi:1… 2.0.0 (…  1.13e7 NL        8 [Tue…    31 1 [Two … NL33 [Zui…
#>  3 ZA5913_3  doi:1… 2.0.0 (…  1.13e7 NL       10 [Thu…    26 1 [Two … NL32 [Noo…
#>  4 ZA5913_4  doi:1… 2.0.0 (…  1.13e7 NL       14 [Mon…    23 1 [Two … NL22 [Gel…
#>  5 ZA5913_5  doi:1… 2.0.0 (…  1.13e7 NL       10 [Thu…    31 2 [Thre… NL33 [Zui…
#>  6 ZA5913_6  doi:1… 2.0.0 (…  1.13e7 NL        8 [Tue…    33 1 [Two … NL41 [Noo…
#>  7 ZA5913_7  doi:1… 2.0.0 (…  1.13e7 NL       10 [Thu…    21 1 [Two … NL13 [Dre…
#>  8 ZA5913_8  doi:1… 2.0.0 (…  1.13e7 NL       10 [Thu…    44 1 [Two … NL13 [Dre…
#>  9 ZA5913_9  doi:1… 2.0.0 (…  1.13e7 NL        5 [Sat…    19 1 [Two … NL34 [Zee…
#> 10 ZA5913_10 doi:1… 2.0.0 (…  1.13e7 NL       14 [Mon…    22 1 [Two … NL22 [Gel…
#> # … with 25 more rows, and 28 more variables: d7 <dbl+lbl>, d8 <dbl+lbl>,
#> #   d25 <dbl+lbl>, d60 <dbl+lbl>, qa10_3 <dbl+lbl>, qa10_2 <dbl+lbl>,
#> #   qa10_1 <dbl+lbl>, qa7_4 <dbl+lbl>, qa7_2 <dbl+lbl>, qa7_3 <dbl+lbl>,
#> #   qa7_1 <dbl+lbl>, qa7_5 <dbl+lbl>, qd3_1 <dbl+lbl>, qd3_2 <dbl+lbl>,
#> #   qd3_3 <dbl+lbl>, qd3_4 <dbl+lbl>, qd3_5 <dbl+lbl>, qd3_6 <dbl+lbl>,
#> #   qd3_7 <dbl+lbl>, qd3_8 <dbl+lbl>, qd3_9 <dbl+lbl>, qd3_10 <dbl+lbl>,
#> #   qd3_11 <dbl+lbl>, qd3_12 <dbl+lbl>, qd3_13 <dbl+lbl>, qd3_14 <dbl+lbl>, …
#>