Subset all surveys in a wave

The function harmonizes the variable names of surveys (of class survey) that are imported from an external file as a wave with with read_surveys.

Usage

subset_surveys(survey_list, rowid = "rowid", subset_names = NULL)

subset_waves(waves, subset_names = NULL)

Arguments

survey_list

A list of surveys imported with read_surveys.

rowid

The unique row (observation) identifier in the files. Defaults to "rowid", which is the default of the importing functions in this package.

subset_names

The names of the variables that should be kept from all surveys in the list that contains the wave of surveys. Defaults to NULL in which case it returns all variables without subsetting.

waves

A list of surveys imported with read_surveys.

Value

The list of surveys with harmonized variable names.

Details

It is likely that you want to harmonize the variable names with harmonize_var_names first.

Examples

examples_dir <- system.file("examples", package = "retroharmonize")
survey_list <- dir(examples_dir)[grepl("\\.rds", dir(examples_dir))]

example_surveys <- read_surveys(
  file.path( examples_dir, survey_list), 
  save_to_rds = FALSE)
metadata <- metadata_surveys_create(example_surveys)

metadata$var_name_suggested <- label_normalize(metadata$var_name)

metadata$var_name_suggested[metadata$label_orig == "age education"] <- "age_education"

hnw <- harmonize_var_names(survey_list = example_surveys, 
                           metadata    = metadata )
                           
subset_surveys (hnw, subset_names = c("uniqid", "w1", "age_education"))
#> [[1]]
#> # A tibble: 35 × 3
#>    rowid       uniqid    w1
#>    <chr>        <dbl> <dbl>
#>  1 ZA5913_1  11339367 0.875
#>  2 ZA5913_2  11339664 0.838
#>  3 ZA5913_3  11339746 1.01 
#>  4 ZA5913_4  11339759 0.476
#>  5 ZA5913_5  11339885 2.09 
#>  6 ZA5913_6  11340290 1.14 
#>  7 ZA5913_7  11340422 0.787
#>  8 ZA5913_8  11340508 0.745
#>  9 ZA5913_9  11340547 0.824
#> 10 ZA5913_10 11340820 0.720
#> # … with 25 more rows
#> 
#> [[2]]
#> # A tibble: 50 × 3
#>    rowid        uniqid    w1
#>    <chr>         <dbl> <dbl>
#>  1 ZA6863_1  110000064 1.49 
#>  2 ZA6863_2  110000417 1.75 
#>  3 ZA6863_3  110000473 1.39 
#>  4 ZA6863_4  110000503 1.25 
#>  5 ZA6863_5  110000663 0.609
#>  6 ZA6863_6  110000889 0.870
#>  7 ZA6863_7  110001276 1.40 
#>  8 ZA6863_8  110001319 0.932
#>  9 ZA6863_9  110001541 1.42 
#> 10 ZA6863_10 110001656 1.12 
#> # … with 40 more rows
#> 
#> [[3]]
#> # A tibble: 45 × 3
#>    rowid        uniqid    w1
#>    <chr>         <dbl> <dbl>
#>  1 ZA7576_1   50000481 0.827
#>  2 ZA7576_2  110000076 0.348
#>  3 ZA7576_3  110000343 0.385
#>  4 ZA7576_4  110000473 0.385
#>  5 ZA7576_5  110000493 0.743
#>  6 ZA7576_6  110000897 0.635
#>  7 ZA7576_7  110001041 0.883
#>  8 ZA7576_8  110001192 0.465
#>  9 ZA7576_9  110001274 1.28 
#> 10 ZA7576_10 110001344 0.909
#> # … with 35 more rows
#>