Subset surveys

This is a wrapper function for various procedures to reduce the size of surveys by removing variables that are not harmonized.

Usage

subset_surveys(
  survey_list,
  survey_paths = NULL,
  rowid = "rowid",
  subset_name = "subset",
  subset_vars = NULL,
  crosswalk_table = NULL,
  import_path = NULL,
  export_path = NULL
)

subset_save_surveys(
  crosswalk_table,
  subset_name = "subset",
  survey_list = NULL,
  survey_paths = NULL,
  import_path = NULL,
  export_path = NULL
)

harmonize_survey_variables(
  crosswalk_table,
  subset_name = "subset",
  survey_list = NULL,
  survey_paths = NULL,
  import_path = NULL,
  export_path = NULL
)

subset_waves(waves, subset_vars = NULL)

Arguments

survey_list

A list of surveys imported with read_surveys. If set to NULL, the survey_path should give full path to the surveys.

survey_paths

A vector of full file paths to the surveys to subset.

rowid

The unique row (observation) identifier in the files. Defaults to "rowid", which is the default of the importing functions in this package.

subset_name

An identifier for the survey subset.

subset_vars

The names of the variables that should be kept from all surveys in the list that contains the wave of surveys. Defaults to NULL in which case it returns all variables without subsetting.

crosswalk_table

A crosswalk table created by crosswalk_table_create or a manually created crosstable including at least filename, var_name_orig, var_name_target and optionally var_label_orig and var_label_target. This parameter is optional and defaults to NULL.

import_path

The path to the survey files. Defaults to NULL.

export_path

The path where the subsets should be saved. Defaults to NULL when the surveys are not saved but returned as a list.

waves

A list of surveys imported with read_surveys.

Value

A list of surveys or save individual rds files on the export_path. The function does not return a value. It saves the subsetted surveys into .rds files.

Details

This function allows several workflows. Subsetting can be based on a vector of variable names given by survey_path, or on the basis of a crosstable. The subset_save_surveys can be called directly.

subset_surveys will also harmonize the variable names if the var_name_target is optionally defined in the crosswalk_table input. harmonize_survey_variables is a wrapper and will require that the new (target) variable names are present in a valid crosstable.

See also

Examples

examples_dir <- system.file("examples", package = "retroharmonize")
survey_list <- dir(examples_dir)[grepl("\\.rds", dir(examples_dir))]

example_surveys <- read_surveys(
  file.path( examples_dir, survey_list), 
  save_to_rds = FALSE)
subset_surveys(survey_list = example_surveys, 
              subset_vars = c("rowid", "isocntry", "qa10_1", "qa14_1"), 
              subset_name = "subset_example")
#> [[1]]
#> # A tibble: 35 × 3
#>    rowid     isocntry                qa10_1
#>    <chr>     <chr>                <dbl+lbl>
#>  1 ZA5913_1  NL       2 [Tend not to trust]
#>  2 ZA5913_2  NL       2 [Tend not to trust]
#>  3 ZA5913_3  NL       3 (NA) [DK]          
#>  4 ZA5913_4  NL       1 [Tend to trust]    
#>  5 ZA5913_5  NL       1 [Tend to trust]    
#>  6 ZA5913_6  NL       1 [Tend to trust]    
#>  7 ZA5913_7  NL       1 [Tend to trust]    
#>  8 ZA5913_8  NL       1 [Tend to trust]    
#>  9 ZA5913_9  NL       2 [Tend not to trust]
#> 10 ZA5913_10 NL       2 [Tend not to trust]
#> # … with 25 more rows
#> 
#> [[2]]
#> # A tibble: 50 × 3
#>    rowid     isocntry                qa14_1
#>    <chr>     <chr>                <dbl+lbl>
#>  1 ZA6863_1  NL       3 [DK]               
#>  2 ZA6863_2  NL       1 [Tend to trust]    
#>  3 ZA6863_3  NL       3 [DK]               
#>  4 ZA6863_4  NL       1 [Tend to trust]    
#>  5 ZA6863_5  NL       1 [Tend to trust]    
#>  6 ZA6863_6  NL       2 [Tend not to trust]
#>  7 ZA6863_7  NL       1 [Tend to trust]    
#>  8 ZA6863_8  NL       3 [DK]               
#>  9 ZA6863_9  NL       1 [Tend to trust]    
#> 10 ZA6863_10 NL       1 [Tend to trust]    
#> # … with 40 more rows
#> 
#> [[3]]
#> # A tibble: 45 × 3
#>    rowid     isocntry                qa14_1
#>    <chr>     <chr>                <dbl+lbl>
#>  1 ZA7576_1  ES       2 [Tend not to trust]
#>  2 ZA7576_2  NL       1 [Tend to trust]    
#>  3 ZA7576_3  NL       1 [Tend to trust]    
#>  4 ZA7576_4  NL       2 [Tend not to trust]
#>  5 ZA7576_5  NL       1 [Tend to trust]    
#>  6 ZA7576_6  NL       1 [Tend to trust]    
#>  7 ZA7576_7  NL       1 [Tend to trust]    
#>  8 ZA7576_8  NL       3 [DK]               
#>  9 ZA7576_9  NL       2 [Tend not to trust]
#> 10 ZA7576_10 NL       2 [Tend not to trust]
#> # … with 35 more rows
#> 
test_survey <- read_rds (
  file = system.file("examples", "ZA7576.rds",
                     package = "retroharmonize")
)

  test_metadata <- metadata_create ( test_survey )
  test_metadata <- test_metadata[c(1,7,18),]
  ctable <- crosswalk_table_create(test_metadata)
#> Joining, by = c("id", "filename", "var_name_orig", "var_name_target", "val_numeric_orig", "val_numeric_target", "val_label_orig", "val_label_target")
#> Joining, by = c("id", "filename", "var_name_orig", "var_name_target", "val_numeric_orig", "val_numeric_target", "val_label_orig", "val_label_target")

subset_save_surveys  ( crosswalk_table = ctable, 
                       subset_name = "tested",
                       survey_list =  test_survey,
                       import_path = NULL
)
#> $ZA7576
#> # A tibble: 45 × 3
#>    rowid     isocntry                qa14_3
#>    <chr>     <chr>                <dbl+lbl>
#>  1 ZA7576_1  ES       2 [Tend not to trust]
#>  2 ZA7576_2  NL       1 [Tend to trust]    
#>  3 ZA7576_3  NL       1 [Tend to trust]    
#>  4 ZA7576_4  NL       3 [DK]               
#>  5 ZA7576_5  NL       1 [Tend to trust]    
#>  6 ZA7576_6  NL       1 [Tend to trust]    
#>  7 ZA7576_7  NL       1 [Tend to trust]    
#>  8 ZA7576_8  NL       1 [Tend to trust]    
#>  9 ZA7576_9  NL       1 [Tend to trust]    
#> 10 ZA7576_10 NL       1 [Tend to trust]    
#> # … with 35 more rows
#> 
test_survey <- read_rds (
  file = system.file("examples", "ZA7576.rds",
                     package = "retroharmonize")
)

  test_metadata <- metadata_create ( test_survey )
  test_metadata <- test_metadata[c(1,7,18),]
  ctable_2 <- crosswalk_table_create(test_metadata)
#> Joining, by = c("id", "filename", "var_name_orig", "var_name_target", "val_numeric_orig", "val_numeric_target", "val_label_orig", "val_label_target")
#> Joining, by = c("id", "filename", "var_name_orig", "var_name_target", "val_numeric_orig", "val_numeric_target", "val_label_orig", "val_label_target")
  ctable_2$var_name_target  <- ifelse(ctable$var_name_orig == "qa14_3", 
                                     "trust_ecb", 
                                      ctable$var_name_orig)

subset_save_surveys  ( crosswalk_table = ctable_2, 
                       subset_name = "tested",
                       survey_list =  test_survey,
                       import_path = NULL
)
#> $ZA7576
#> # A tibble: 45 × 3
#>    rowid     isocntry             trust_ecb
#>    <chr>     <chr>                <dbl+lbl>
#>  1 ZA7576_1  ES       2 [Tend not to trust]
#>  2 ZA7576_2  NL       1 [Tend to trust]    
#>  3 ZA7576_3  NL       1 [Tend to trust]    
#>  4 ZA7576_4  NL       3 [DK]               
#>  5 ZA7576_5  NL       1 [Tend to trust]    
#>  6 ZA7576_6  NL       1 [Tend to trust]    
#>  7 ZA7576_7  NL       1 [Tend to trust]    
#>  8 ZA7576_8  NL       1 [Tend to trust]    
#>  9 ZA7576_9  NL       1 [Tend to trust]    
#> 10 ZA7576_10 NL       1 [Tend to trust]    
#> # … with 35 more rows
#>